do not install Mediawiki twice

+++ mediawiki-html/entrypoint.sh
@@ -9,10 +9,14 @@ make_mediawiki_directory () {
 }
 
 install_mediawiki () {
-    wget https://releases.wikimedia.org/mediawiki/1.39/mediawiki-1.39.6.tar.gz
-    tar -xvf mediawiki-*.tar.gz
-    mv mediawiki-*/* /var/www/${MEDIAWIKI_HOST}/html
-    rm -r mediawiki-*
+    if ! cat /var/www/${MEDIAWIKI_HOST}/URL | grep -q https://releases.wikimedia.org/mediawiki/1.39/mediawiki-1.39.6.tar.gz; then
+        rm -r /var/www/${MEDIAWIKI_HOST}/html
+        wget https://releases.wikimedia.org/mediawiki/1.39/mediawiki-1.39.6.tar.gz
+        tar -xvf mediawiki-*.tar.gz
+        mv mediawiki-*/* /var/www/${MEDIAWIKI_HOST}/html
+        rm -r mediawiki-*
+        echo https://releases.wikimedia.org/mediawiki/1.39/mediawiki-1.39.6.tar.gz > /var/www/${MEDIAWIKI_HOST}/URL
+    fi
 }
 
 make_nginx_own_mediawiki () {
